MethodMassage-Comopatibility is no longer required to build a VMMaker image. Ken's code to fix github.com/OpenSmalltalk/opensmalltalk-vm/issues/741

+++ platforms/Cross/plugins/UnicodePlugin/UnicodePlugin.h
@@ -4,10 +4,10 @@
 int unicodeClipboardGet(unsigned short *utf16, int utf16Length);
 void unicodeClipboardPut(unsigned short *utf16, int utf16Length);
 int unicodeClipboardSize(void);
-void unicodeDrawString(char *utf8, int utf8Length, int *wPtr, int *hPtr, unsigned int *bitmapPtr);
+void unicodeDrawString(char *utf8, int utf8Length, long *wPtr, long *hPtr, unsigned int *bitmapPtr);
 int unicodeGetFontList(char *str, int strLength);
 int unicodeGetXRanges(char *utf8, int utf8Length, int *resultPtr, int resultLength);
-void unicodeMeasureString(char *utf8, int utf8Length, int *wPtr, int *hPtr);
+void unicodeMeasureString(char *utf8, int utf8Length, long *wPtr, long *hPtr);
 void unicodeSetColors(int fgRed, int fgGreen, int fgBlue, int bgRed, int bgGreen, int bgBlue, int mapBGToTransparent);
 void unicodeSetFont(char *fontName, int fontSize, int boldFlag, int italicFlag, int antiAliasFlag);
